Fixed issue in get_transformation script, and moved script into lfric_common to remove code duplication.
