+++ run_configs/lfric/gungho.py
@@ -4,6 +4,10 @@
 #  For further details please refer to the file COPYRIGHT
 #  which you should have received as part of this distribution
 # ##############################################################################
+'''
+A simple build script for gungho_model
+'''
+
 import logging
 
 from fab.build_config import BuildConfig
@@ -18,29 +22,10 @@ from fab.steps.psyclone import psyclone, preprocess_x90
 from fab.tools import ToolBox
 
 from grab_lfric import lfric_source_config, gpl_utils_source_config
-from lfric_common import configurator, fparser_workaround_stop_concatenation
+from lfric_common import (configurator, fparser_workaround_stop_concatenation,
+                          get_transformation_script)
 
 logger = logging.getLogger('fab')
-
-
-def get_transformation_script(fpath, config):
-    ''':returns: the transformation script to be used by PSyclone.
-    :rtype: Path
-
-    '''
-    optimisation_path = config.source_root / 'optimisation' / 'meto-spice'
-    for base_path in [config.source_root, config.build_output]:
-        try:
-            relative_path = fpath.relative_to(base_path)
-        except ValueError:
-            pass
-    local_transformation_script = optimisation_path / (relative_path.with_suffix('.py'))
-    if local_transformation_script.exists():
-        return local_transformation_script
-    global_transformation_script = optimisation_path / 'global.py'
-    if global_transformation_script.exists():
-        return global_transformation_script
-    return ""
 
 
 if __name__ == '__main__':
